add a retry strategy based on the error provided by bitbucket
bitbucket provides a list of default reviewers that can be rejected when a pull request is created. Added a strategy to remove invalid reviewers and retry the PR creation. This will specificaly work in the scenario where the creator of the pull request is also part of the default reviewer. It is most likely to happen if you are debugging the application and using your credential on a repository for which you are a default reviewer.